fix(desktop): split lib modules under size guard
Move focused desktop entrypoint helpers out of lib.rs so the file stays well below the default size guard instead of relying on a one-line exception. Drop the lib.rs override now that the entrypoint is back under the shared limit.
